+++ openquake/hazardlib/source/point.py
@@ -26,8 +26,8 @@ from openquake.hazardlib.geo.surface.planar import (
     build_planar, PlanarSurface, planin_dt, get_rupdims)
 from openquake.hazardlib.pmf import PMF
 from openquake.hazardlib.scalerel.point import PointMSR
-from openquake.hazardlib.source.base import (ParametricSeismicSource,
-                                             MagDepAspectRatio)
+from openquake.hazardlib.aspect_ratio import MagDepAspectRatio
+from openquake.hazardlib.source.base import ParametricSeismicSource
 from openquake.hazardlib.source.rupture import (
     ParametricProbabilisticRupture)
 from openquake.hazardlib.geo.utils import get_bounding_box, angular_distance
