+++ openquake/hazardlib/sourceconverter.py
@@ -27,7 +27,8 @@ from openquake.baselib.node import context, striptag, Node, node_to_dict
 from openquake.hazardlib import geo, mfd, pmf, source, tom, valid, InvalidFile
 from openquake.hazardlib.tom import PoissonTOM
 from openquake.hazardlib.source_group import SourceGroup
-from openquake.hazardlib.source.base import MagDepAspectRatio
+from openquake.hazardlib.aspect_ratio import (
+    MagDepAspectRatio, get_aspect_ratio, build_aspect_ratio_node)
 from openquake.hazardlib.source.multi_fault import MultiFaultSource
 
 U32 = numpy.uint32
@@ -45,50 +46,6 @@ KNOWN_MFDS = ('incrementalMFD', 'truncGutenbergRichterMFD',
 EXCLUDE_FROM_GEOM_PROPS = (
     'Polygon', 'Point', 'MultiPoint', 'LineString', '3D MultiLineString',
     '3D MultiPolygon', 'posList')
-
-
-def get_aspect_ratio(node):
-    """
-    Get the aspect ratio. It can be either a single float value or a
-    piecewise function which is evaluated during rupture generation.
-    """
-    try:
-        arf = node.aspectRatioFunction
-    except AttributeError:
-        return ~node.ruptAspectRatio
-
-    func_type = ~arf.type
-
-    if func_type == 'linear_piecewise':
-        npoints = len(arf.mag_points)
-        if npoints != 2:
-            raise ValueError(
-                f"Should be two elements [(MinAR, MinMmin), (MaxAR, Mmax)] "
-                f"in linear_piecewise kind of ruptAspectRatio ({npoints})")
-        points = [(float(p['mag']), float(p['aratio'])) for p in arf.mag_points]
-        if any(m <= 0 for m, _ in points):
-            raise ValueError(
-                f"aspectRatioFunction magnitudes must be positive: "
-                f"{[m for m, _ in points]}")
-        if points[0][0] >= points[1][0]:
-            raise ValueError(
-                f"aspectRatioFunction points must be in ascending magnitude "
-                f"order: mag[0]={points[0][0]} >= mag[1]={points[1][0]}")
-        return MagDepAspectRatio("linear_piecewise", points)
-
-    raise ValueError(f"Unsupported aspectRatioFunction type: {func_type}")
-
-
-def build_aspect_ratio_node(rar):
-    if not isinstance(rar, MagDepAspectRatio):
-        return Node("ruptAspectRatio", text=rar)
-    if rar.func_type == "linear_piecewise":
-        point_nodes = [Node("mag_point", {"mag": m, "aratio": a})
-                       for m, a in rar.mag_points]
-        return Node("aspectRatioFunction", nodes=[
-            Node("type", text="linear_piecewise"),
-            Node("mag_points", nodes=point_nodes)])
-    raise ValueError(f"Unsupported aspectRatioFunction type: {rar.func_type}")
 
 
 def extract_dupl(values):
